Week8/python/practice/planning_pattern_agentic/requirements.txt
langchain
langgraph
langchain-groq
streamlit
fpdf
python-dotenv
logging
